services/auth/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
prometheus-client==0.20.0
pydantic==2.6.3
